prepare finish all tomorrow

+++ exercises/conversions/using_as.rs
@@ -1,3 +1,10 @@
+/*
+ * @Author: ss
+ * @Date: 2023-10-19 19:50:06
+ * @LastEditTime: 2023-10-19 19:50:06
+ * @Description: 
+ * @FilePath: \rust-rustlings-2023-autumn-Ss-shuang123\exercises\conversions\using_as.rs
+ */
 // using_as.rs
 //
 // Type casting in Rust is done via the usage of the `as` operator. Please note
@@ -9,12 +16,11 @@
 //
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
-
-// I AM NOT DONE
+ 
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / (values.len() as f64)
 }
 
 fn main() {
